# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Lab Insight Workspace API
  version: 0.1.0
  description: Versioned contract for ingestion, schema review, and A2UI-streamed analysis.
  license:
    name: Proprietary demo
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
security: []
paths:
  /health:
    get:
      summary: Read service health
      operationId: getHealth
      responses:
        "200":
          description: Service health
          content:
            application/json:
              schema: {$ref: "#/components/schemas/HealthResponse"}
  /demo/initialize:
    post:
      summary: Initialize deterministic demo data
      operationId: initializeDemo
      responses:
        "200":
          description: Demo data initialized
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DemoInitializeResponse"}
  /demo/files/{file_name}:
    get:
      summary: Download a bundled original demo source
      operationId: downloadDemoFile
      parameters:
        - {name: file_name, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Original demo source file
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
  /ingestions:
    get:
      summary: List ingestion batches
      operationId: listIngestions
      responses:
        "200":
          description: Ingestion batches
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/IngestionBatch"}}
                  total: {type: integer}
  /ingestions/files:
    post:
      summary: Upload a structured data file
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
                vendor_hint: {type: [string, "null"]}
      responses:
        "202":
          description: File accepted
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IngestionBatch"}
  /ingestions/text:
    post:
      summary: Submit unstructured source text
      operationId: ingestText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_name, content]
              properties:
                source_name: {type: string}
                content: {type: string, minLength: 1, maxLength: 50000}
                vendor_hint: {type: [string, "null"]}
      responses:
        "202":
          description: Text accepted
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IngestionBatch"}
  /ingestions/{batch_id}:
    get:
      summary: Read an ingestion batch
      operationId: getIngestion
      parameters:
        - {name: batch_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Batch detail
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IngestionBatch"}
  /ingestions/{batch_id}/mapping:
    get:
      summary: Read a mapping draft
      operationId: getMapping
      parameters:
        - {name: batch_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Mapping draft
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MappingDraft"}
    put:
      summary: Update a mapping draft
      operationId: updateMapping
      parameters:
        - {name: batch_id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MappingDraft"}
      responses:
        "200":
          description: Updated mapping
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MappingDraft"}
  /ingestions/{batch_id}/profile:
    get:
      summary: Read the source data profile
      operationId: getIngestionProfile
      parameters:
        - {name: batch_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Column-level profile for review
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DataProfile"}
  /ingestions/{batch_id}/preview:
    get:
      summary: Preview original source rows
      operationId: getIngestionPreview
      parameters:
        - {name: batch_id, in: path, required: true, schema: {type: string}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        "200":
          description: Bounded source preview
          content:
            application/json:
              schema:
                type: object
                required: [rows, row_count, limit]
                properties:
                  rows: {type: array, items: {type: object, additionalProperties: true}}
                  row_count: {type: integer}
                  limit: {type: integer}
  /ingestions/{batch_id}/commit:
    post:
      summary: Publish an ingestion batch
      operationId: commitIngestion
      parameters:
        - {name: batch_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Published batch
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IngestionBatch"}
  /conversations:
    post:
      summary: Create an analysis conversation
      operationId: createConversation
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Conversation"}
  /conversations/{conversation_id}/messages/stream:
    post:
      summary: Stream an A2UI analysis response
      operationId: streamMessage
      parameters:
        - {name: conversation_id, in: path, required: true, schema: {type: string}}
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string}}
        - {name: Last-Event-ID, in: header, required: false, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateMessageRequest"}
      responses:
        "200":
          description: Ordered SSE events whose data fields are A2UI v0.9.1 envelopes
          content:
            text/event-stream:
              schema: {type: string}
  /conversations/{conversation_id}/messages:
    get:
      summary: List persisted conversation messages
      operationId: listMessages
      parameters:
        - {name: conversation_id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Conversation history with completed A2UI snapshots
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/Message"}}
                  total: {type: integer}
  /conversations/{conversation_id}/messages/{message_id}/cancel:
    post:
      summary: Cancel an active response stream
      operationId: cancelMessage
      parameters:
        - {name: conversation_id, in: path, required: true, schema: {type: string}}
        - {name: message_id, in: path, required: true, schema: {type: string}}
      responses:
        "202":
          description: Cancellation accepted
          content:
            application/json:
              schema:
                type: object
                required: [message_id, status]
                properties:
                  message_id: {type: string}
                  status: {type: string}
  /conversations/{conversation_id}/a2ui/actions:
    post:
      summary: Handle a standard A2UI surface action
      operationId: handleA2UIAction
      parameters:
        - {name: conversation_id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/A2UIActionRequest"}
      responses:
        "202":
          description: Action accepted
          content:
            application/json:
              schema: {type: object}
  /analysis/suggestions:
    get:
      summary: List suggested analysis questions
      operationId: getSuggestions
      responses:
        "200":
          description: Suggested questions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {type: object}}
  /assets/{asset_name}:
    get:
      summary: Read a validated same-origin chart asset
      operationId: getAsset
      parameters:
        - {name: asset_name, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Cached chart image
          content:
            image/png: {schema: {type: string, format: binary}}
            image/jpeg: {schema: {type: string, format: binary}}
components:
  schemas:
    HealthResponse:
      type: object
      required: [status, version, database, model, visualization_mcp]
      properties:
        status: {type: string, const: ok}
        version: {type: string}
        database: {type: object}
        model: {type: object}
        visualization_mcp: {type: object}
    DemoInitializeResponse:
      type: object
      required: [initialized, record_count, batch_ids]
      properties:
        initialized: {type: boolean}
        record_count: {type: integer}
        batch_ids: {type: array, items: {type: string}}
    IngestionBatch:
      type: object
      required: [batch_id, source_type, source_name, status, record_count, created_at, updated_at]
      properties:
        batch_id: {type: string}
        source_type: {type: string, enum: [CSV, XLSX, TEXT]}
        source_name: {type: string}
        vendor_hint: {type: [string, "null"]}
        status: {type: string, enum: [UPLOADED, PROFILING, MAPPING, NEEDS_REVIEW, COMMITTING, READY, FAILED]}
        record_count: {type: integer}
        current_stage: {type: [string, "null"]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        download_url: {type: [string, "null"]}
    FieldMapping:
      type: object
      required: [source_field, target_field, confidence, transform, reason, status]
      properties:
        source_field: {type: string}
        target_field: {type: [string, "null"]}
        confidence: {type: number, minimum: 0, maximum: 1}
        transform: {type: string}
        reason: {type: string}
        status: {type: string, enum: [SUGGESTED, CONFIRMED, MODIFIED, IGNORED]}
        sample_before: {type: array}
        sample_after: {type: array}
        warnings: {type: array, items: {type: string}}
    MappingDraft:
      type: object
      required: [batch_id, version, mappings, missing_required_fields, can_commit]
      properties:
        batch_id: {type: string}
        version: {type: integer}
        mappings: {type: array, items: {$ref: "#/components/schemas/FieldMapping"}}
        missing_required_fields: {type: array, items: {type: string}}
        can_commit: {type: boolean}
    DataProfile:
      type: object
      required: [row_count, column_count, columns, warnings]
      properties:
        row_count: {type: integer}
        column_count: {type: integer}
        warnings: {type: array, items: {type: string}}
        columns:
          type: array
          items:
            type: object
            required: [name, inferred_type, null_rate, distinct_count, sample_values, warnings]
            properties:
              name: {type: string}
              inferred_type: {type: string}
              null_rate: {type: number, minimum: 0, maximum: 1}
              distinct_count: {type: integer}
              sample_values: {type: array}
              warnings: {type: array, items: {type: string}}
    Conversation:
      type: object
      required: [conversation_id, title, created_at, updated_at]
      properties:
        conversation_id: {type: string}
        title: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CreateMessageRequest:
      type: object
      required: [question, a2uiClientCapabilities]
      properties:
        question: {type: string, minLength: 1, maxLength: 1000}
        reasoningEnabled:
          type: boolean
          default: true
          description: Controls provider-native thinking for planning, visualization tool selection, and final answer generation.
        filters: {type: object}
        a2uiClientCapabilities:
          type: object
          required: [supportedCatalogIds]
          properties:
            supportedCatalogIds:
              type: array
              items: {type: string}
    Message:
      type: object
      required: [message_id, role, content, status, created_at]
      properties:
        message_id: {type: string}
        role: {type: string, enum: [USER, ASSISTANT]}
        content: {type: string}
        status: {type: string, enum: [STREAMING, COMPLETED, CANCELLED, FAILED]}
        a2ui_surface_snapshot: {type: [object, "null"]}
        created_at: {type: string, format: date-time}
        completed_at: {type: [string, "null"], format: date-time}
    A2UIActionRequest:
      type: object
      required: [surface_id, action_id, name]
      properties:
        surface_id: {type: string}
        action_id: {type: string}
        name: {type: string}
        context: {type: object}